+++ arch/x86/core/crt0.S
@@ -314,13 +314,13 @@ __csSet:
 
 	call	_x86_data_copy
 
-#ifdef CONFIG_APP_SHARED_MEM
+#ifdef CONFIG_USERSPACE
 	movl	$_app_smem_start, %edi /* DATA in RAM (dest) */
 	movl	$_app_smem_rom_start, %esi /* DATA in ROM (src) */
 	movl	$_app_smem_num_words, %ecx /* Size of DATA in quad bytes */
 
 	call	_x86_data_copy
-#endif /* CONFIG_APP_SHARED_MEM */
+#endif /* CONFIG_USERSPACE */
 
 #endif /* CONFIG_XIP */
