+++ arch/x86/core/x86_mmu.c
@@ -24,7 +24,7 @@ MMU_BOOT_REGION((u32_t)&_image_rodata_start, (u32_t)&_image_rodata_size,
 		MMU_ENTRY_READ | MMU_ENTRY_USER |
 		MMU_ENTRY_EXECUTE_DISABLE);
 
-#ifdef CONFIG_APP_SHARED_MEM
+#ifdef CONFIG_USERSPACE
 MMU_BOOT_REGION((u32_t)&_app_smem_start, (u32_t)&_app_smem_size,
 		MMU_ENTRY_WRITE | MMU_ENTRY_RUNTIME_USER |
 		MMU_ENTRY_EXECUTE_DISABLE);
