+++ include/app_memory/app_memdomain.h
@@ -5,7 +5,7 @@
 #include <misc/dlist.h>
 #include <kernel.h>
 
-#ifdef CONFIG_APP_SHARED_MEM
+#ifdef CONFIG_USERSPACE
 
 #if defined(CONFIG_X86)
 #define MEM_DOMAIN_ALIGN_SIZE _STACK_BASE_ALIGN
@@ -104,5 +104,5 @@ struct z_app_region {
 #define K_APP_BMEM_SECTION(ptn) .bss
 #define K_APPMEM_PARTITION_DEFINE(name)
 
-#endif /* CONFIG_APP_SHARED_MEM */
+#endif /* CONFIG_USERSPACE */
 #endif /* ZEPHYR_INCLUDE_APP_MEMORY_APP_MEMDOMAIN_H_ */
