+++ include/arch/x86/linker.ld
@@ -207,7 +207,7 @@ SECTIONS
 	__gcov_bss_size = __gcov_bss_end - __gcov_bss_start;
 #endif /* CONFIG_COVERAGE_GCOV */
 
-#ifdef CONFIG_APP_SHARED_MEM
+#ifdef CONFIG_USERSPACE
 	/* APP SHARED MEMORY REGION */
 #define SMEM_PARTITION_ALIGN(size) MMU_PAGE_ALIGN
 #define APP_SHARED_ALIGN  MMU_PAGE_ALIGN
@@ -219,12 +219,12 @@ SECTIONS
 	_app_smem_num_words = _app_smem_size >> 2;
 	_app_smem_rom_start = LOADADDR(_APP_SMEM_SECTION_NAME);
 	_app_smem_num_words = _app_smem_size >> 2;
-#endif /* CONFIG_APP_SHARED_MEM */
+#endif /* CONFIG_USERSPACE */
 
 	SECTION_PROLOGUE(_BSS_SECTION_NAME, (NOLOAD OPTIONAL),)
 	{
 	MMU_PAGE_ALIGN
-#if !defined(CONFIG_APP_SHARED_MEM)
+#if !defined(CONFIG_USERSPACE)
 	_image_ram_start = .;
 #endif
 	/*
