+++ lib/libc/minimal/source/stdlib/malloc.c
@@ -16,16 +16,16 @@
 #include <logging/log.h>
 LOG_MODULE_DECLARE(os);
 
-#ifdef CONFIG_APP_SHARED_MEM
+#ifdef CONFIG_USERSPACE
 K_APPMEM_PARTITION_DEFINE(z_malloc_partition);
 #endif
 
 #if (CONFIG_MINIMAL_LIBC_MALLOC_ARENA_SIZE > 0)
-#ifdef CONFIG_APP_SHARED_MEM
+#ifdef CONFIG_USERSPACE
 #define POOL_SECTION K_APP_DMEM_SECTION(z_malloc_partition)
 #else
 #define POOL_SECTION .data
-#endif /* CONFIG_APP_SHARED_MEM */
+#endif /* CONFIG_USERSPACE */
 
 K_MUTEX_DEFINE(malloc_mutex);
 SYS_MEM_POOL_DEFINE(z_malloc_mem_pool, &malloc_mutex, 16,
