+++ samples/userspace/shared_mem/prj.conf
@@ -1,2 +1,1 @@
 CONFIG_USERSPACE=y
-CONFIG_APP_SHARED_MEM=y
