--- subsys/app_memory/CMakeLists.txt
@@ -1,2 +1,0 @@
-zephyr_library()
-zephyr_library_sources(app_memdomain.c)
