+++ subsys/testsuite/ztest/include/ztest_test.h
@@ -143,7 +143,7 @@ static inline void unit_test_noop(void)
  */
 
 /* definitions for use with testing application shared memory   */
-#ifdef CONFIG_APP_SHARED_MEM
+#ifdef CONFIG_USERSPACE
 #define ZTEST_DMEM	K_APP_DMEM(ztest_mem_partition)
 #define ZTEST_BMEM	K_APP_BMEM(ztest_mem_partition)
 #define ZTEST_SECTION	K_APP_DMEM_SECTION(ztest_mem_partition)
