+++ tests/kernel/threads/thread_init/src/main.c
@@ -207,7 +207,7 @@ void test_main(void)
 {
 	k_thread_access_grant(k_current_get(), &thread_preempt, &stack_preempt,
 			      &start_sema, &end_sema);
-#ifdef CONFIG_APP_SHARED_MEM
+#ifdef CONFIG_USERSPACE
 	k_mem_domain_add_thread(&ztest_mem_domain, T_KDEFINE_COOP_THREAD);
 	k_mem_domain_add_thread(&ztest_mem_domain, T_KDEFINE_PREEMPT_THREAD);
 #endif
